Updated tests
tests should work now
